backend/requirements.txt
# Open NotebookLM — Python dependencies
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
httpx>=0.27.0
pydantic>=2.9.0

# Document parsing
pdfplumber>=0.11.0
python-docx>=1.1.0
python-pptx>=0.6.23
trafilatura>=1.12.0

# RAG
chromadb>=0.5.0
sentence-transformers>=3.0.0
transformers>=4.44.0,<4.47
accelerate>=0.33.0

# TTS: Piper (system binary) + Silero (torch.hub)
numpy<2
torch>=2.0,<2.2
torchaudio>=2.0,<2.2
omegaconf>=2.0

# Audio
Pillow>=10.0.0
faster-whisper>=1.1.1
vosk>=0.3.45

# Config
pyyaml>=6.0

# Multipart uploads
python-multipart>=0.0.9
